--- a/src/lib/Hydra/Controller/Channel.pm
+++ b/src/lib/Hydra/Controller/Channel.pm
@@ -22,8 +22,7 @@ sub channel : Chained('/') PathPart('channel/custom') CaptureArgs(3) {
 
     my $lastSuccessful = $c->model('DB::Builds')->find(
         { 'eval.hasnewbuilds' => 1
-        , project => $projectName
-        , jobset => $jobsetName
+        , jobset_id => $c->stash->{jobset}->id,
         , job => $channelName
         , buildstatus => 0
         },