@tomberek 2021-06-01
@grahamc it builds and runs, but my current hydra doesn't have enough jobs to measure the perf difference.