@grahamc 2021-06-01
https://hydra.nixos.org/project/nixpkgs is loading in like 2.5s now :), down from 5-25s.